Fix wrong path to book module to preserve data
